% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylog.R
\name{mylog}
\alias{mylog}
\title{Computing the log}
\usage{
mylog(x)
}
\arguments{
\item{x}{a vector if length n}
}
\value{
a vector of length n with the log of x
}
\description{
Computing the log
}
\examples{
v <- c(1, 2, 3)
mylog(v)

}
